Add files via upload
these pictures are the mean of the Fide rating of chess players as it is observable it has been experiencing a downward trend from the first to the end but the result is that the number of fide players has been increasing noticeably. New players are usually new to the field and perform below the average rating.