Add files via upload
Kernel density of All types, Standard, rapid, and blitz to understand the different perform of the players.
By examining graph A-1, it can be seen that the kernel density of the international rating of female players with an international rating below 1,400 has increased a lot after COVID-19 compared to before, this means that the proportion of semi-professional players with an international rank below 1,400 Compared to international degrees, it has increased a lot. The highest kernel density difference in international grade is 1,200. For men, this increase in kernel density continues up to the international rating of 1,600, which is the highest kernel density difference in the international rating of 1,200. But this difference is less in women than in men. In fact, more new player-arrived women have entered the sample than men which might be a result of the queen gambit mini-series which the main role was a woman who becomes an international chess champion for the first time in a male-dominated chess environment. Chess has also grown and the number of observations in the game in the last month of April 2023 is 413,945, which is about 2 times the first month of February 2015, which was 200,033. The same trend can be seen in rapid and blitz games. The number of observations in the game of the last month of rapid is 261,622, which is about 6 times that of the first month of data (45,310). In the game, the last month of blitz is 182,879, which is about 5 times the first month of data (37,071). In fact, the number of chess enthusiasts has increased, and these people who are newcomers often have an international rank below 1,700.
On the other hand, for women with an international degree above 1,400, the density of the international degree after COVID-19 is much lower than before COVID-19, this difference decreases as the international degree increases. In men, with an international grade above 1,600, this decreasing trend can be seen similarly. The reason for the decreasing trend is several factors; The first reason is that there are fewer professional people, which is completely normal, and the second reason is the low impact of factors such as COVID-19 on the performance of these professional players, after the international rank of 2,500, which is the international rank for the great masters, the number of these people is very small, and these people are the most professional players. continuously In all the graphs of Appendix A, this distinction between the density of players with an international rank above 2,500 is almost zero(the other kernel density graphs are available in ).