The Impact of Time Constraints on Gender Performance Gap Evidence from Chess Player
This study investigates the impact of COVID-19 and time pressure as anxiety stimuli on the performance of male and female chess players. The research population consists of 445,941 players from 198 different federations of the World Chess Federation from February 2015 to April 2023. Chess players compete in three types of games: standard, rapid, and blitz, with the only difference being the time constraint for completing the game. The performance of chess players is measured by their international rating in all three types of games and evaluated using an unbalanced pooled aggregate least squares model. In this model, month and year are controlled as seasonal effects, and the impact of COVID-induced anxiety conditions on the international rating of players is examined. The results indicate that the average international rating has decreased after the COVID period for both genders. However, this decrease is not solely attributed to COVID; additionally, attached graphs show a significant increase in the number of non-professional players with ratings below the average, contributing to the decrease in the average international rating post-COVID. Controlling for factors such as gender, age, and dummy variables, the model results show that COVID has also led to an increase in the international rating of players. This increase is more pronounced in standard games, which have the least time pressure, and least in blitz games, which have the highest time pressure. With increasing age, the international rating of players has increased in all three types of competitions. This increase is higher in blitz and rapid games, where participants are younger compared to standard games. Women, who have, on average, lower age and international ratings compared to men, experienced a smaller decrease in the average international rating during the COVID period compared to men and, according to the model results, have experienced a greater increase in international rating compared to men. The increase in the international rating of women is highest in rapid games and lowest in standard games and is insignificant. No relationship was found between increased time pressure and a greater decrease in female performance compared to males in COVID-era games. Women have, on average, weaker performance compared to men and withdraw from chess games sooner.